medibot-flask/requirements.txt
streamlit
langchain
numpy
langchain-core
langchain-community
langchain-huggingface
huggingface-hub
sentence-transformers
faiss-cpu
python-dotenv
torch==2.1.0+cpu
torchvision==0.16.0+cpu
torchaudio==2.1.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu
